+++ release.sh
@@ -10,7 +10,7 @@ declare -a unixOS=("osx-x64" "linux-x64")
 
 for rid in "${winOS[@]}"
 do
-	dotnet publish ./FakeServer/FakeServer.csproj -c release -r $rid
+	dotnet publish ./FakeServer/FakeServer.csproj -c release -r $rid /p:PackAsTool=false
 	cd ./FakeServer/bin/release/netcoreapp2.2/$rid/publish/
 	zip -r ../../../../../../releases/fakeserver-$rid.zip .
 	cd ../../../../../../
@@ -18,7 +18,7 @@ done
 
 for rid in "${unixOS[@]}"
 do
-	dotnet publish ./FakeServer/FakeServer.csproj -c release -r $rid
+	dotnet publish ./FakeServer/FakeServer.csproj -c release -r $rid /p:PackAsTool=false
 	cd ./FakeServer/bin/release/netcoreapp2.2/$rid/publish/
 	tar -cvzf ../../../../../../releases/fakeserver-$rid.tar.gz *
 	cd ../../../../../../