+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * ADDED: Token endpoint to support username/password in JSON content
 * ADDED: Define stored data id-field name
 * CHANGED: Rename Common property from appsettings.json to DataStore
+* CHANGED: Remove authentication.json and add Authentication settings to appsettings.json
 
 ### [0.9.1] - 2019-02-08
 * CHANGED: Target framework to .NET Core 2.2